Gios Api consumer module refactored
