+++ api/package.json
@@ -15,6 +15,8 @@
     "express": "^4.17.1",
     "jsonwebtoken": "^8.5.1",
     "mongoose": "^5.9.10",
-    "nodemailer": "^6.4.6"
+    "node-schedule": "^1.3.2",
+    "nodemailer": "^6.4.6",
+    "superagent": "^5.2.2"
   }
 }
